% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.window.R
\name{cal.window}
\alias{cal.window}
\title{Calendar time cut of survival data}
\usage{
cal.window(dat, window, surv = c("enter", "exit", "event", "birthdate"))
}
\arguments{
\item{dat}{Input data frame. Must contain survival data and a birth date.}

\item{window}{Vector of length two; the time interval}

\item{surv}{Vector of length four giving the names of the central variables
in 'dat'.}
}
\value{
A data frame of the same form as the input data frame, but 'cut' as
desired. Intervals exceeding \code{window[2]} will be given \code{event = 0}
}
\description{
For a given time interval, each spell is cut so that it fully lies in the
given time interval
}
\details{
The \code{window} must be in the order \code{(begin, end)}
}
\examples{

dat <- data.frame(enter = 0, exit = 5.731, event = 1,
birthdate = 1962.505, x = 2)
window <- c(1963, 1965)
dat.trim <- cal.window(dat, window)  

}
\seealso{
\code{\link{age.window}}, \code{\link{coxreg}},
\code{\link{aftreg}}
}
\author{
Göran Broström
}
\keyword{survival}
