% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eha-defunct}
\alias{eha-defunct}
\alias{ghq}
\alias{glmmboot}
\alias{glmmbootFit}
\alias{glmmML}
\alias{glmmML.fit}
\title{Defunct functions}
\usage{
ghq(...)

glmmboot(...)

glmmbootFit(...)

glmmML(...)

glmmML.fit(...)
}
\arguments{
\item{...}{input parameters}
}
\description{
These functions were duplicates of functions in the package \pkg{glmmML}.
}
\details{
Instead of using these functions, use the corresponding functions in 
\pkg{glmmML} with the same name
}
