% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha-package.R
\docType{package}
\name{eha}
\alias{eha}
\alias{eha-package}
\title{eha}
\description{
Sampling of risk sets in Cox regression, selections in 
the Lexis diagram, bootstrapping. Parametric proportional 
hazards fitting with left truncation and right censoring for 
common families of distributions, piecewise constant hazards, 
and discrete models. AFT regression for left truncated and 
right censored data. Binary and Poisson regression for 
clustered data, fixed and random effects with bootstrapping.
}
