% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghq.R
\name{ghq}
\alias{ghq}
\title{Gauss-Hermite}
\usage{
ghq(n.points = 1, modified = TRUE)
}
\arguments{
\item{n.points}{Number of points.}

\item{modified}{Multiply by exp(zeros**2)? Default is TRUE.}
}
\value{
A list vith components \item{zeros}{The zeros (abscissas).}
\item{weights}{The weights}
}
\description{
Calculates the zeros and weights needed for Gauss-Hermite quadrature.
}
\details{
Based on a Fortran 66 subroutine written by professor Jianming Jin.
}
\note{
The code is modified to suit the purpose of glmmML, with the
permission of professor Jin.
}
\examples{

ghq(15, FALSE)

}
\references{
Gauss-Hermite
}
\seealso{
\code{\link{glmmML}}
}
\author{
Jianming Jin, Univ. of Illinois, Urbana-Campaign
}
\keyword{math}
