% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmboot.R
\name{glmmboot}
\alias{glmmboot}
\title{Generalized Linear Models with fixed effects grouping}
\usage{
glmmboot(formula, family = binomial, data, cluster, weights, subset,
  na.action, offset, start.coef = NULL, control = list(epsilon = 1e-08,
  maxit = 200, trace = FALSE), boot = 0)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. The details of
model specification are given below.}

\item{family}{Currently, the only valid values are \code{binomial} and
\code{poisson}. The binomial family allows for the \code{logit} and
\code{cloglog} links.}

\item{data}{an optional data frame containing the variables in the model.
By default the variables are taken from `environment(formula)', typically
the environment from which `glmmML' is called.}

\item{cluster}{Factor indicating which items are correlated.}

\item{weights}{Case weights.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{See glm.}

\item{offset}{this can be used to specify an a priori known component to be
included in the linear predictor during fitting.}

\item{start.coef}{starting values for the parameters in the linear
predictor.  Defaults to zero.}

\item{control}{Controls the convergence criteria. See
\code{\link{glm.control}} for details.}

\item{boot}{number of bootstrap replicates. If equal to zero, no test of
significance of the grouping factor is performed.}
}
\value{
The return value is a list, an object of class 'glmmboot'.
\item{coefficients}{Estimated regression coefficients} \item{logLik}{the max
log likelihood} \item{cluster.null.deviance}{Deviance without the
clustering} \item{frail}{The estimated cluster effects} \item{bootLog}{The
logLik values from the bootstrap samples} \item{bootP}{Bootstrap p value}
\item{variance}{Variance covariance matrix} \item{sd}{Standard error of
regression parameters} \item{boot_rep}{No. of bootstrap replicates}
\item{mixed}{Logical} \item{deviance}{Deviance} \item{df.residual}{Its
degrees of freedom} \item{aic}{AIC} \item{boot}{Logical} \item{call}{The
function call}
}
\description{
Fits grouped GLMs with fixed group effects. The significance of the grouping
is tested by simulation, with a bootstrap approach.
}
\details{
The simulation is performed by simulating new response vectors from the
fitted probabilities without clustering, and comparing the maximized log
likelihoods. The maximizations are performed by profiling out the grouping
factor. It is a very fast procedure, compared to \code{\link{glm}}, when the
grouping factor has many levels.
}
\note{
There is no overall intercept for this model; each cluster has its own
intercept. See \code{frail}
}
\examples{

## Not run:
id <- factor(rep(1:20, rep(5, 20)))
y <- rbinom(100, prob = rep(runif(20), rep(5, 20)), size = 1)
x <- rnorm(100)
dat <- data.frame(y = y, x = x, id = id)
res <- glmmboot(y ~ x, cluster = id, data = dat, boot = 500)
## End(Not run)
##system.time(res.glm <- glm(y ~ x + id, family = binomial))

}
\seealso{
\code{link{glmmML}}, \code{\link{optim}}, \code{lmer} in the
package \code{lme4}, and \code{glmmPQL} in the package \code{MASS}.
}
\author{
Göran Broström
}
\keyword{nonlinear}
\keyword{regression}
