% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazards.R
\name{hazards}
\alias{hazards}
\title{Get baseline hazards atoms from fits from}
\usage{
hazards(x, cum = TRUE, ...)
}
\arguments{
\item{x}{A \code{reg} object.}

\item{cum}{Logical: Should the cumulative hazards be returned?}

\item{...}{Additional arguments for various methods.}
}
\value{
A list where each component is a two-column matrix representing
hazard atoms from one stratum. The first column is event time, and the 
second column is the corresponding hazard atom.
}
\description{
Get baseline hazards atoms from fits from
}
