% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisco.R
\name{HiscoHisclass}
\alias{HiscoHisclass}
\title{HISCO to HISCLASS transformation}
\usage{
HiscoHisclass(hisco, status = NULL, relation = NULL, urban = NULL,
debug = FALSE)
}
\arguments{
\item{hisco}{Hisco codes to be transformed to hisclass.}

\item{status}{Optional standard description of status.}

\item{relation}{Relation between owner of occupation and self.}

\item{urban}{Logical, "Is residence in an urban area?"}

\item{debug}{Logical, prints intermediate values if TRUE.}
}
\value{
A vector of hisclass codes, same length as input hisco.
}
\description{
HISCO to HISCLASS transformation
}
\references{
Van Leeuwen, M. and Maas, I. (2011). HISCLASS. A Historical
International Social Class Scheme. Leuwen University Press.
}
\author{
Göran Broström with translation and modification of a Stata do.
}
