% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{import_strata}
\alias{import_strata}
\alias{strata}
\title{\code{strata} function imported from \code{survival}}
\description{
This function is imported from the \code{survival} package. See
\code{\link[survival]{strata}}.
}
