% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{import_Surv}
\alias{import_Surv}
\alias{Surv}
\title{\code{Surv} function imported from \code{survival}}
\description{
This function is imported from the \code{survival} package. See
\code{\link[survival]{Surv}}.
}
