% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe.R
\name{oe}
\alias{oe}
\title{Create an oe object}
\usage{
oe(count, exposure)
}
\arguments{
\item{count}{Number of events, a non-negative integer-valued vector.}

\item{exposure}{exposure time corresponding to count. 
A positive numeric vector.}
}
\description{
Create an \emph{oe} ("occurrence/exposure") object, used as a response 
variable in a model formula specifically in \code{\link{tpchreg}}.
}
\seealso{
\code{\link{tpchreg}}.
}
