% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aftGompreg.R
\name{print.aftGompreg}
\alias{print.aftGompreg}
\title{Prints aftGompreg objects}
\usage{
\method{print}{aftGompreg}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{A \code{aftreg} object}

\item{digits}{Precision in printing}

\item{\dots}{Not used.}
}
\value{
No value is returned.
}
\description{
The hazard, the cumulative hazard, the density, and the survivor baseline
functions are plotted.
}
\note{
Doesn't work for threeway or higher order interactions. Use
\code{\link[survival]{print.coxph}} in that case.
}
\seealso{
\code{\link{phreg}}, \code{\link[survival]{print.coxph}}
}
\author{
Göran Broström
}
\keyword{regression}
\keyword{survival}
