% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.glmmML.R
\name{print.glmmML}
\alias{print.glmmML}
\title{Prints a 'glmmML' object.}
\usage{
\method{print}{glmmML}(x, digits = max(3, getOption("digits") - 3),
  na.print = "", ...)
}
\arguments{
\item{x}{The glmmML object}

\item{digits}{Number of printed digits.}

\item{na.print}{How to print NAs}

\item{\dots}{Additional parameters, which are ignored.}
}
\value{
A short summary of the object is printed.
}
\description{
A glmmML object is the output of \code{glmmML}.
}
\details{
Nothing in particular.
}
\note{
This is the only summary method available for the moment.
}
\seealso{
\code{\link{glmmML}}
}
\author{
Göran Broström
}
\keyword{print}
