% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.logrank.R
\name{print.logrank}
\alias{print.logrank}
\title{Prints logrank objects}
\usage{
\method{print}{logrank}(x, digits = max(options()$digits - 4, 6), ...)
}
\arguments{
\item{x}{A \code{logrank} object}

\item{digits}{Precision in printing}

\item{\dots}{Not used.}
}
\value{
The input is returned invisibly.
}
\description{
The result of \code{logrank} is printed
}
\seealso{
\code{\link{logrank}}, \code{\link{coxreg}}
}
\author{
Göran Broström
}
\keyword{regression}
\keyword{survival}
