% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.aftGompreg.R
\name{print.summary.aftGompreg}
\alias{print.summary.aftGompreg}
\title{Prints summary.aftGompreg objects}
\usage{
\method{print}{summary.aftGompreg}(x, digits = max(getOption("digits") - 3, 3), short = FALSE, ...)
}
\arguments{
\item{x}{A \code{summary.aftreg} object, typically the result of running
\code{summary.aftreg}, summary on a phreg object.}

\item{digits}{Output format.}

\item{short}{Logical: If TRUE, short output, only regression.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Prints summary.aftGompreg objects
}
\seealso{
\code{\link{aftreg}}, \code{\link{summary.aftreg}}
}
\author{
Göran Broström
}
\keyword{survival}
