% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.coxreg.R
\name{print.summary.coxreg}
\alias{print.summary.coxreg}
\title{Prints summary.coxreg objects}
\usage{
\method{print}{summary.coxreg}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{summary.coxreg} object, typically the result of running
\code{summary.coxreg}, summary on a coxreg object.}

\item{digits}{Output format.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Prints summary.coxreg objects
}
\seealso{
\code{\link{coxreg}}, \code{\link{summary.coxreg}}
}
\author{
Göran Broström
}
\keyword{survival}
