% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.phreg.R
\name{print.summary.phreg}
\alias{print.summary.phreg}
\title{Prints summary.phreg objects}
\usage{
\method{print}{summary.phreg}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{A \code{summary.phreg} object, typically the result of running
\code{summary.phreg}, summary on a phreg object.}

\item{digits}{Output format.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Prints summary.phreg objects
}
\seealso{
\code{\link{phreg}}, \code{\link{summary.phreg}}
}
\author{
Göran Broström
}
\keyword{survival}
