% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.tpchreg.R
\name{print.summary.tpchreg}
\alias{print.summary.tpchreg}
\title{Prints summary.tpchreg objects}
\usage{
\method{print}{summary.tpchreg}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{A \code{summary.tpchreg} object, typically the result of running
\code{summary.tpchreg}, summary on a tpchreg object.}

\item{digits}{Output format.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Prints summary.tpchreg objects
}
\seealso{
\code{\link{tpchreg}}, \code{\link{summary.tpchreg}}
}
\author{
Göran Broström
}
\keyword{survival}
