% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tpchreg.R
\name{print.tpchreg}
\alias{print.tpchreg}
\title{Prints tpchreg objects}
\usage{
\method{print}{tpchreg}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{A \code{tpchreg} object, typically the result of running
\code{tpchreg}}

\item{digits}{Output format.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
More "pretty-printing"
}
\details{
Doesn't work with three-way or higher interactions.
}
\seealso{
\code{\link{tpchreg}}, \code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{survival}
