% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_nageklerke.aftreg.R
\name{r2_nagelkerke.aftreg}
\alias{r2_nagelkerke.aftreg}
\title{Calculate Nagelkerke's pseudo-R2 for a \code{\link[eha]{aftreg}} object}
\usage{
r2_nagelkerke.aftreg(model, ...)
}
\arguments{
\item{model}{A \code{\link[eha]{aftreg}} object}

\item{\dots}{Additional ...}
}
\description{
Calculate Nagelkerke's pseudo-R2 for a \code{\link[eha]{aftreg}} object
}
\details{
Based on \code{\link[performance]{r2_nagelkerke}}
}
\examples{
data(mort)
r2_nagelkerke.aftreg(aftreg(Surv(enter, exit, event) ~ ses, param = 'lifeExp', data = mort))
}
\keyword{aftreg}
\keyword{nagelkerke}
\keyword{r2}
