% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtable.R
\name{regtable}
\alias{regtable}
\title{Retrieves regression tables}
\usage{
regtable(x, digits = 3, short = TRUE, ...)
}
\arguments{
\item{x}{A \code{summary.XXreg} object, typically the result of running
\code{summary.XXreg}, summary on a XXreg object.}

\item{digits}{Output format.}

\item{short}{If TRUE, return only coefficients table.}

\item{\dots}{Other arguments.}
}
\value{
A character data frame, ready to print in various formats.
}
\description{
Retrieves regression tables
}
\note{
Should not be used if interactions present.
}
\seealso{
\code{\link{coxreg}}, \code{\link{summary.coxreg}}
}
\author{
Göran Broström
}
\keyword{survival}
