% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsGomp.R
\name{rmsGomp}
\alias{rmsGomp}
\title{Restricted mean survival calculation for the Gompertz distribution by
numerical integration of the survival function.}
\usage{
rmsGomp(
  from = 0,
  to = Inf,
  scale = 1,
  shape = 1,
  param = c("canonical", "rate"),
  fit
)
}
\arguments{
\item{from}{Start time for the integration}

\item{to}{Stop time of the integration.}

\item{scale}{The scale parameter.}

\item{shape}{The shape parameter.}

\item{param}{Which parametrization? Defaults to 'canonical', the other choice
is 'rate' (not developed yet, avoid until next version of eha!).}

\item{fit}{The result of a call to 'aftGompreg'. If not missing, scale and 
shape will be taken from the fit, and corresponding argumens in the call will 
be ignored.}
}
\description{
Restricted mean survival calculation for the Gompertz distribution by
numerical integration of the survival function.
}
\examples{
res <- aftGompreg(Surv(enter-60, exit-60, event) ~ ses.50 + strata(sex), data = oldmort)
rmsGomp(from = 0, to = 40, fit = res)
}
