% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftGompreg.R
\name{summary.aftGompreg}
\alias{summary.aftGompreg}
\title{Creates a summary of aftGompreg objects}
\usage{
\method{summary}{aftGompreg}(object, ...)
}
\arguments{
\item{object}{A \code{aftGompreg} object}

\item{\dots}{Additional ...}
}
\description{
Creates a summary of aftGompreg objects
}
\examples{

## The function is currently defined as
function (object, ...) 
print(object)

}
\seealso{
\code{\link{summary.aftreg}}
}
\author{
Göran Broström
}
\keyword{summary}
\keyword{survival}
