% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftreg.R
\name{summary.aftreg}
\alias{summary.aftreg}
\title{Prints aftreg objects}
\usage{
\method{summary}{aftreg}(object, ..., use.drop1 = TRUE)
}
\arguments{
\item{object}{A \code{aftreg} object}

\item{\dots}{Additional ...}

\item{use.drop1}{logical; Use \code{\link[stats]{drop1}} to compute `p` values based on log odds etc.}
}
\description{
Prints aftreg objects
}
\examples{

## The function is currently defined as
function (object, ...)
print(object)

}
\seealso{
\code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
