\name{summary.coxreg}
\alias{summary.coxreg}

\title{Prints coxreg objects}
\description{
This is the same as \code{\link{print.coxreg}}
}
\usage{
summary.coxreg(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A \code{coxreg} object}
  \item{\dots}{Additional ...}
}

\author{Göran Broström}

\seealso{\code{\link{print.coxreg}}}
\examples{
## The function is currently defined as
function (object, ...) 
print(object)
}
\keyword{survival}
\keyword{print}
