% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glmmboot.R
\name{summary.glmmboot}
\alias{summary.glmmboot}
\title{Summary of a glmmboot object}
\usage{
\method{summary}{glmmboot}(object, ...)
}
\arguments{
\item{object}{A glmmboot object}

\item{\dots}{Additional arguments}
}
\value{
Nothing is returned.
}
\description{
It simply calls \code{print.glmmboot}
}
\details{
A summary method will be written soon.
}
\note{
Preliminary
}
\seealso{
\code{\link{print.glmmboot}}
}
\author{
Göran Broström
}
\keyword{print}
