% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glmmML.R
\name{summary.glmmML}
\alias{summary.glmmML}
\title{Summary of a glmmML object}
\usage{
\method{summary}{glmmML}(object, ...)
}
\arguments{
\item{object}{A glmmML object}

\item{\dots}{Additional arguments}
}
\value{
Nothing is returned.
}
\description{
It simply calls \code{print.glmmML}
}
\note{
Preliminary
}
\seealso{
\code{\link{print.glmmML}}
}
\author{
Göran Broström
}
\keyword{print}
