\name{summary.phreg}
\alias{summary.phreg}

\title{Prints phreg objects}
\description{
This is the same as \code{\link{print.phreg}}
}
\usage{
\method{summary}{phreg}(object, \ldots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A \code{phreg} object}
  \item{\dots}{Additional ...}
}

\author{\enc{Göran Broström}{George Bridgewater}}

\seealso{\code{\link{print.coxreg}}}
\examples{
## The function is currently defined as
function (object, ...) 
print(object)
}
\keyword{survival}
\keyword{print}
