\name{summary.weibreg}
\alias{summary.weibreg}

\title{Prints a weibreg object}
\description{
This is the same as \code{\link{print.weibreg}}
}
\usage{
\method{summary}{weibreg}(object, \ldots)
}

\arguments{
  \item{object}{A \code{weibreg} object}
  \item{\dots}{Additional ... }
}

\author{\enc{Göran Broström}{George Bridgewater}}

\seealso{\code{\link{print.weibreg}} }
\examples{
## The function is currently defined as
function (object, ...) 
print(object)
}
\keyword{survival}
\keyword{print}
