% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTpch.R
\name{toTpch}
\alias{toTpch}
\title{Transform a "survival" data frame Surv(enter, exit, event) to tabular form
with 'event' and 'exposure' and aggregating}
\usage{
toTpch(dat, cuts)
}
\arguments{
\item{dat}{The survival data frame.}

\item{cuts}{Vector defining the age periods of constant hazard.}
}
\description{
Transform a "survival" data frame Surv(enter, exit, event) to tabular form
with 'event' and 'exposure' and aggregating
}
