% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha.R
\docType{package}
\name{eha-package}
\alias{eha}
\alias{eha-package}
\title{eha: Event History Analysis}
\description{
Parametric proportional hazards fitting with left truncation and right censoring for common families of distributions, piecewise constant hazards, and discrete models. Parametric accelerated failure time models for left truncated and right censored data. Proportional hazards models for tabular and register data. Sampling of risk sets in Cox regression, selections in the Lexis diagram, bootstrapping. Broström (2022) \doi{10.1201/9780429503764}.
}
\details{
Eha enhances the recommended \pkg{survival} package in several ways, 
see the description. The main applications in mind are demography and 
epidemiology. For standard Cox regression analysis the function 
\code{\link[survival]{coxph}} in \pkg{survival} is still recommended. The function 
\code{\link{coxreg}} in \pkg{eha} in fact calls coxph for the standard kind 
of analyses.
}
\references{
Broström, G. (2012). \emph{Event History Analysis with R}, Chapman and Hall/CRC 
Press, Boca Raton, FL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ehar.se/r/eha/}
  \item Report bugs at \url{https://github.com/goranbrostrom/eha/issues}
}

}
\author{
\strong{Maintainer}: Göran Broström \email{goran.brostrom@umu.se}

Other contributors:
\itemize{
  \item Jianming Jin [contributor]
  \item Altan Orhon \email{altan@uw.edu} [contributor]
}

}
