% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frail.fit.R
\name{frail.fit}
\alias{frail.fit}
\title{Frailty experiment}
\usage{
frail.fit(X, Y, rs, strats, offset, init, max.survs, frailty, control)
}
\arguments{
\item{X}{design matrix}

\item{Y}{survival object}

\item{rs}{output from \code{\link{risksets}}}

\item{strats}{strata}

\item{offset}{offset}

\item{init}{start values}

\item{max.survs}{for sampling of riskset survivors}

\item{frailty}{grouping variable}

\item{control}{control of optimization}
}
\description{
Utilizing GLMM models: Experimental, not exported (yet).
}
