% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geome.fit.R
\name{geome.fit}
\alias{geome.fit}
\title{Constant intensity discrete time proportional hazards}
\usage{
geome.fit(X, Y, rs, strats, offset, init, max.survs, method = "ml",
  control)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{Survival object}

\item{rs}{risk set produced by \code{\link{risksets}}}

\item{strats}{Stratum indicator}

\item{offset}{Offset}

\item{init}{Initial values}

\item{max.survs}{Maximal survivors}

\item{method}{"ml", always, i.e., this argument is ignored.}

\item{control}{See \code{coxreg}.}
}
\value{
See the code.
}
\description{
This function is called from \code{coxreg}. A user may call it directly.
}
\note{
Nothing special

\code{coxreg} is a defunct function
}
\references{
See \code{\link{coxreg}}.
}
\seealso{
\code{\link{coxreg}}
}
\author{
Göran Broström
}
\keyword{survival}
