% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank.R
\name{logrank}
\alias{logrank}
\title{The Log-rank test}
\usage{
logrank(formula, data = parent.frame())
}
\arguments{
\item{formula}{formula a formula object, with the response on the left of a ~
operator, and the terms on the right. The response must be a survival object
as returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the
formula.}
}
\value{
A list of class \code{logrank} with components
\item{test.statistic}{The logrank (score) test statistic.}
\item{df}{The degrees of freedom of the test statistic.}
\item{p.value}{The p value of the test.}
\item{hazards}{A list of two-column matrices, describing event times and 
corresponding hazard atoms in each stratum (class 'hazdata').}
}
\description{
Performs the log-rank test on survival data, possibly stratified.
}
\note{
The test is performed by fitting a Cox regression model and reporting
its \code{score test}. With tied data, this might be slightly different from
the true logrank test, but the difference is unimprtant in practice.
}
\examples{
fit <- logrank(Surv(enter, exit, event) ~ sex, data = oldmort)
fit
}
\seealso{
\code{\link{coxreg}}, \code{\link{print.logrank}}.
}
\author{
Göran Broström
}
\keyword{survival}
