% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coxreg.R
\name{print.coxreg}
\alias{print.coxreg}
\title{Prints coxreg objects}
\usage{
\method{print}{coxreg}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{A \code{coxreg} object, typically the result of running
\code{coxreg}}

\item{digits}{Output format.}

\item{\dots}{Other arguments.}
}
\value{
No value is returned.
}
\description{
More "pretty-printing" than \code{print.coxph}, which is a fall-back for
'difficult' objects.
}
\details{
Doesn't work with three-way and higher interactions, in which case
\code{print.coxph} is used. Prints also output from \code{\link{mlreg}}.
}
\seealso{
\code{\link{coxreg}}, \code{\link[survival]{print.coxph}}
}
\author{
Göran Broström
}
\keyword{survival}
