% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.risksets.R
\name{print.risksets}
\alias{print.risksets}
\title{Prints a summary of the content of a set of risk sets.}
\usage{
\method{print}{risksets}(x, ...)
}
\arguments{
\item{x}{An object of class 'risksets'.}

\item{...}{Not used for the moment.}
}
\value{
No value is returned; the function prints summary statistics of risk
sets.
}
\description{
Given the output from \code{risksets}, summary statistics are given for it.
}
\note{
There is no \code{summary.risksets} yet. On the TODO list.
}
\examples{

rs <- with(mort, risksets(Surv(enter, exit, event)))
print(rs)

}
\seealso{
\code{risksets}
}
\author{
Göran Broström
}
\keyword{summary}
