\name{summary.aftreg}
\alias{summary.aftreg}

\title{Prints aftreg objects}
\description{
This is the same as \code{\link{print.aftreg}}
}
\usage{
\method{summary}{aftreg}(object, \ldots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A \code{aftreg} object}
  \item{\dots}{Additional ...}
}

\author{\enc{Göran Broström}{George Bridgewater}}

\seealso{\code{\link{print.coxreg}}}
\examples{
## The function is currently defined as
function (object, ...) 
print(object)
}
\keyword{survival}
\keyword{print}
