% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftreg.R
\name{summary.aftreg}
\alias{summary.aftreg}
\title{Prints aftreg objects}
\usage{
\method{summary}{aftreg}(object, ...)
}
\arguments{
\item{object}{A \code{aftreg} object}

\item{\dots}{Additional ...}
}
\description{
Prints aftreg objects
}
\examples{

## The function is currently defined as
function (object, ...) 
print(object)

}
\seealso{
\code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
