% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coxreg.R
\name{summary.coxreg}
\alias{summary.coxreg}
\title{A summary of coxreg objects.}
\usage{
\method{summary}{coxreg}(object, ...)
}
\arguments{
\item{object}{A \code{coxreg} object}

\item{\dots}{Additional ...}
}
\description{
A summary of coxreg objects.
}
\examples{

fit <- coxreg(Surv(enter, exit, event) ~ sex + civ, data = oldmort)
summary(fit)
 
}
\seealso{
\code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
