% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coxreg.R
\name{summary.coxreg}
\alias{summary.coxreg}
\title{Prints coxreg objects}
\usage{
\method{summary}{coxreg}(object, ...)
}
\arguments{
\item{object}{A \code{coxreg} object}

\item{\dots}{Additional ...}
}
\description{
This is the same as \code{\link{print.coxreg}}
}
\examples{

## The function is currently defined as
function (object, ...) 
print(object)

}
\seealso{
\code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
