% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phreg.R
\name{summary.phreg}
\alias{summary.phreg}
\title{Prints phreg objects}
\usage{
\method{summary}{phreg}(object, ...)
}
\arguments{
\item{object}{A \code{phreg} object}

\item{\dots}{Additional ...}
}
\description{
This is the same as \code{\link{print.phreg}}
}
\examples{

## The function is currently defined as
function (object, ...) 
print(object)

}
\seealso{
\code{\link{print.coxreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
