% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phreg.R
\name{summary.phreg}
\alias{summary.phreg}
\title{A summary of phreg objects.}
\usage{
\method{summary}{phreg}(object, ...)
}
\arguments{
\item{object}{A \code{phreg} object}

\item{\dots}{Additional ...}
}
\description{
A summary of phreg objects.
}
\examples{

fit <- phreg(Surv(enter, exit, event) ~ sex + civ, 
data = oldmort[oldmort$region == "town", ])
summary(fit)
 
}
\seealso{
\code{\link{print.phreg}}
}
\author{
Göran Broström
}
\keyword{survival}
