% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tpchreg.R
\name{summary.tpchreg}
\alias{summary.tpchreg}
\title{Summary for tpchreg objects}
\usage{
\method{summary}{tpchreg}(object, ...)
}
\arguments{
\item{object}{A \code{tpchreg} object.}

\item{\dots}{Additional ...}
}
\description{
Summary for tpchreg objects
}
\examples{

## The function is currently defined as
## function (object, ...) 

}
\seealso{
\code{\link{tpchreg}}
}
\author{
Göran Broström
}
\keyword{summary}
\keyword{survival}
