% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.weibreg.R
\name{summary.weibreg}
\alias{summary.weibreg}
\title{Prints a weibreg object}
\usage{
\method{summary}{weibreg}(object, ...)
}
\arguments{
\item{object}{A \code{weibreg} object}

\item{\dots}{Additional ...}
}
\description{
This is the same as \code{\link{print.weibreg}}
}
\examples{

## The function is currently defined as
function (object, ...) 
print(object)

}
\seealso{
\code{\link{print.weibreg}}
}
\author{
Göran Broström
}
\keyword{print}
\keyword{survival}
