% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDate.R
\name{toDate}
\alias{toDate}
\title{Convert time in years since "0000-01-01" to a date.}
\usage{
toDate(times)
}
\arguments{
\item{times}{a vector of durations}
}
\value{
A vector of dates as character strings of the type "1897-05-21".
}
\description{
This function uses \code{as.Date} and a simple linear transformation.
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
toDate(1897.357)

}
\seealso{
\code{\link{toTime}}
}
\author{
Göran Broström
}
\keyword{survival}
