% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTime.R
\name{toTime}
\alias{toTime}
\title{Calculate duration in years from "0000-01-01" to a given date}
\usage{
toTime(dates)
}
\arguments{
\item{dates}{A vector of dates in character form or of class \code{Date}}
}
\value{
A vector of durations, as decribed above.
}
\description{
Given a vector of dates, the output is a vector of durations in years since
"0000-01-01".
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
toTime(c("1897-05-16", "1901-11-21"))

}
\seealso{
\code{\link{toDate}}
}
\author{
Göran Broström
}
\keyword{survival}
