% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpchreg.R
\name{tpchreg}
\alias{tpchreg}
\title{Proportional hazards regression with piecewise constant hazards and tabular data.}
\usage{
tpchreg(formula, cuts, data)
}
\arguments{
\item{formula}{a formula}

\item{cuts}{intervals for constant hazard.}

\item{data}{a data frame with event, exposure, age plus covariates}
}
\description{
Proportional hazards regression with piecewise constant hazards and tabular data.
}
